Add map selection
Now we keep two lists: one of all maps (setup) and one of those
in play. Maps have an inPlay field, and are copied from allMaps
to playMaps whenever there is a change / selection starts.
This has some duplication, but allows for easy resets and state
tracking as before. One change would be a minimum on the
number of maps (7?).