Factor out smaller updates, remove sync Cmd
A number of the synchronous Cmd calls (AdvancePhase etc.) were
removed in favour of piping into sub-updates. Still not super
clean, and the names get unwieldy, but apparently sync Cmd
generation with "identity" is a bit of an anti-pattern. I still
like it in the case of SetRandom being called at the end of the
appropriate pick phase. All of those that were factored out could
be manual actions from the user. We shall see.