Refactor app into sections
Not necessary, but it is more neat than a 1k line file. There is
still a consideration of separating general views, setup/play
views, as well as setup/play updates. Maybe past this point is
where I would do that, but it is unlikely there's more screens
to be added.